refactor: improve modularity by separate card and col components from matches
