+++ front/src/pages/comps/matches/index.jsx
@@ -1,80 +1,6 @@
-import { Box, Grid, Stack, Paper, Chip, Typography, Divider, Fab } from '@mui/material';
-import { styled } from '@mui/material/styles';
-import { ExpandLess as ExpandLessIcon, ExpandMore as ExpandMoreIcon } from '@mui/icons-material';
+import { Box, Grid, Typography, Divider } from '@mui/material';
 
-/**
- * Style for item Cards
- */
-const ItemText = styled(Paper)(({ theme }) => ({
-  ...theme.typography.h6,
-  padding: theme.spacing(1),
-  color: theme.palette.text.primary,
-}));
-
-/**
- * Return Grid item for a Card representing of an item
- * @param {mobillsItemObject} param0
- * @param {*} colorMoney
- * @returns
- */
-function itemCard({
-  title, category, date, amount,
-}, colorMoney) {
-  return (
-    <Grid item xs={12} sx={{ padding: 1 }}>
-      <Paper>
-        <Grid container spacing={1}>
-          <Grid item xs={12}>
-            <Stack justifyContent="space-around" direction="row">
-              <Chip label={date} />
-              <Chip label={category} />
-              <Chip label={`R$ ${(amount / 100).toFixed(2)}`} color={colorMoney} />
-            </Stack>
-          </Grid>
-          <Grid item xs={12} sx={{ textAlign: 'center' }}><ItemText elevation={0}>{title}</ItemText></Grid>
-        </Grid>
-      </Paper>
-    </Grid>
-  );
-}
-
-/**
- * Return column of items
- * @param {*} param0
- * @returns
- */
-function ItemCol({ data, colorMoney, amountCardsToShow, setRowProp }) {
-  const arrayData = Object.keys(data);
-  const expanded = amountCardsToShow > process.env.REACT_APP_AMOUNT_COLLAPSED_CARD;
-
-  return (
-    <Grid container item xs={6} sx={{ maxWidth: '350px' }} direction='row' sepacing={1} columns={1}>
-      {arrayData.slice(0, amountCardsToShow).map((item, idx) => itemCard(data[item], colorMoney))}
-      <Grid container item alignItems='center' justifyContent='center'>
-        {arrayData.length > process.env.REACT_APP_AMOUNT_COLLAPSED_CARD ?
-          (
-            expanded ?
-              <Fab size='small' onClick={() => {
-                setRowProp('amountCardsToShow', process.env.REACT_APP_AMOUNT_COLLAPSED_CARD);
-              }}
-              >
-                <ExpandLessIcon />
-              </Fab>
-              :
-              <Fab size='small' onClick={() => {
-                setRowProp('amountCardsToShow', arrayData.length);
-              }}
-              >
-                <ExpandMoreIcon />
-              </Fab>
-          )
-          :
-          null
-        }
-      </Grid>
-    </Grid>
-  );
-}
+import ItemCol from './ItemCol';
 
 /**
  * Return row of items
